requirements.txt
torch>=1.13
torchvision>=0.14
numpy
pandas
scikit-learn
matplotlib
opencv-python
tqdm
Pillow
seaborn


# optional (for Grad-CAM)
torchcam